RunwaySim/training/yolo/requirements.txt
# YOLOv8 학습을 위한 필수 패키지들
ultralytics>=8.0.0
torch>=1.11.0
torchvision>=0.12.0
torchaudio>=0.11.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
pyyaml>=6.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
tensorboard>=2.8.0
tqdm>=4.62.0
psutil>=5.8.0

# 추가 유틸리티
scikit-learn>=1.0.0
scipy>=1.7.0
plotly>=5.0.0 